requirements.txt
pytest
selene==2.0.0rc9
selenium==4.15.2
webdriver-manager
allure-pytest